move github config methods
